@@ -1,4 +1,4 @@
-class AddSecretAndMigrationColumnToJiraIssue < ActiveRecord::Migration
+class AddSecretAndMigrationColumnToJiraIssue < ActiveRecord::Migration[4.2]
   def self.up
     add_column :jira_issues, :secrets_modified, :text, :limit => 255
     add_column :jira_issues, :long_running_migration, :text, :limit => 255
