@@ -1,4 +1,4 @@
-class PreserveMergedJiraIssues < ActiveRecord::Migration
+class PreserveMergedJiraIssues < ActiveRecord::Migration[4.2]
   def self.up
     add_column :jira_issues_and_pushes, :merged, :boolean, :null => false, :default => false
   end
