@@ -1,4 +1,4 @@
-class AddNoJiraToCommitsAndPushes < ActiveRecord::Migration
+class AddNoJiraToCommitsAndPushes < ActiveRecord::Migration[4.2]
   def self.up
     add_column :commits_and_pushes, :no_jira, :boolean, default: false, null: false
   end
