@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class AddEmailSentToPushes < ActiveRecord::Migration
+class AddEmailSentToPushes < ActiveRecord::Migration[4.2]
   def self.up
     add_column :pushes, :email_sent, :boolean, :default => false
   end
