@@ -17,9 +17,9 @@ class JiraIssue < ActiveRecord::Base
   validates :key, uniqueness: { message: 'Keys must be globally unique' }
   validates :key, format: { with: /\A.+-[0-9]+\z/ }
 
-  belongs_to :assignee, class_name: User, inverse_of: :commits, required: false, null: true
-  belongs_to :parent_issue, class_name: JiraIssue, inverse_of: :sub_tasks, required: false, null: true
-  has_many :sub_tasks, class_name: JiraIssue, dependent: :nullify
+  belongs_to :assignee, class_name: 'User', inverse_of: :commits, required: false, null: true
+  belongs_to :parent_issue, class_name: 'JiraIssue', inverse_of: :sub_tasks, required: false, null: true
+  has_many :sub_tasks, class_name: 'JiraIssue', dependent: :nullify
   has_many :commits, foreign_key: 'jira_issue_id', dependent: :nullify
   has_many :jira_issues_and_pushes, class_name: :JiraIssuesAndPushes, inverse_of: :jira_issue, dependent: :destroy
   has_many :pushes, through: :jira_issues_and_pushes
@@ -41,7 +41,7 @@ class JiraIssue < ActiveRecord::Base
   end
 
   def jira_url_for_issue
-    "#{Rails.application.secrets.jira['site']}/browse/#{key}"
+    "#{Rails.application.secrets.jira[:site]}/browse/#{key}"
   end
 
   class << self
