@@ -5,13 +5,13 @@ describe Api::Callbacks::JiraController, type: :controller do
 
   describe 'POST #push' do
     it 'returns success if valid JSON' do
-      post :hook, load_fixture_file('jira_hook_payload.json')
+      post :hook, body: load_fixture_file('jira_hook_payload.json')
       expect(response).to have_http_status(200)
       expect(Delayed::Job.count).to eq(1)
     end
 
     it 'returns bad request if not valid JSON' do
-      post :hook, 'This is not JSON'
+      post :hook, body: 'This is not JSON'
       expect(response).to have_http_status(400)
       expect(Delayed::Job.count).to eq(0)
     end
